Moved get_machine_identifier temporiarly till sorted where it will be and what
